requirements.txt
Django
djangorestframework
kornia
kornia-rs
kornia_moons
opencv-python
drf-yasg
Pillow
numpy